user.pyとusers.pyの結合
